Make default type double.
Match the ompx test out of the box.

+++ benchmarks/sort/cuda/cuda_perf_sort.cu
@@ -12,8 +12,8 @@
 #include <stdlib.h>
 #include <string>
 
-using TYPE = int;
-// using TYPE=double;
+//using TYPE = int;
+ using TYPE=double;
 
 unsigned int seed = 0;
